JPA modification to stop direct use of JpaRepository

+++ src/main/java/com/admiral/tsp/timesheetportal/review/services/ReviewJpaRepo.java
@@ -2,6 +2,7 @@ package com.admiral.tsp.timesheetportal.review.services;
 
 import com.admiral.tsp.timesheetportal.review.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -11,7 +12,8 @@ import java.util.Optional;
 public interface ReviewJpaRepo {
 
     Review makeReview(Review item);
-    Optional<Review> getByID(Long ID);
+    Optional<Review> getByID(Integer ID);
     List<Review> getAll();
+
 
 }
